feat: Add non UB AR + Residual + Norm + Quant fusion
