+++ tests/unittest/_torch/multi_gpu/test_user_buffers.py
@@ -457,10 +457,10 @@ def run_single_rank_ub_pass(
             output_fused = model_opt(input)
         # 3 AR_NORM fusion happens first
         # 2 AR_NORM fused with Quant
-        # 1 AR_NORM replacement
+        # 3 AR_NORM replacement
         # 3 Scaled MM Prologue
         # 2 UB Finalize Removal
-        assert backend.match_count == [3, 0, 2, 0, 1, 0, 3, 0, 2, 0]
+        assert backend.match_count == [3, 0, 2, 0, 3, 0, 3, 0, 2, 0]
         torch.cuda.synchronize()
 
         if rank == 0:
@@ -1013,10 +1013,10 @@ def run_single_rank_ub_pass_fp4(
 
         # 3 AR_NORM fusion happens first
         # 2 AR_NORM fused with Quant
-        # 1 AR_NORM replacement
+        # 3 AR_NORM replacement
         # 3 Scaled MM Prologue
         # 2 UB Finalize Removal
-        assert backend.match_count == [3, 0, 2, 0, 1, 0, 3, 0, 2, 0]
+        assert backend.match_count == [3, 0, 2, 0, 3, 0, 3, 0, 2, 0]
         torch.cuda.synchronize()
         torch.testing.assert_close(output_fused,
                                    output_ref,
